Call out related teams for libtest-json
I'd like to vet the design by experimenting with integration within
Cargo.
I had assumed the T-cargo members of T-testing-devex would be sufficient
but the T-cargo wanted visibility in this, so I'm adding them.

Since I was added T-cargo, I decided to add T-libs-api.  We are working
under T-libs-api and this ends with a stabilization report for them but
figure if I was adding more, I'd include them as well.