fix goal template magic values to show up on the web
Some rows use html comments to mark special tokens to be replaced, but
they only show up in the markdown source, not on the website.

Some people are copying the template from the website, so the rows are
missing from the PRs.

We encode the values in the markdown, so they show up on the web, and add a note
to make sure to use the unencoded values if copying from source.