Fix tests, a CI issue, a template bug, and clean stuff up (#312)
* fix tests

* fix formatting

* remove unused dependencies

* fix comments

* remove async

* move update generation to main crate

* remove unneeded crates

* reformat workspace key

* fix CI: tracking issue id validation

The template asks people to leave this field blank, but CI validates
it's present and valid.

* fix goal template magic values to show up on the web

Some rows use html comments to mark special tokens to be replaced, but
they only show up in the markdown source, not on the website.

Some people are copying the template from the website, so the rows are
missing from the PRs.

We encode the values in the markdown, so they show up on the web, and add a note
to make sure to use the unencoded values if copying from source.

* add milestone validation for rpg updates

this will help people who pass the path to the ./src/ milestone.

* make goal tracking issues mandatory for accepted goals

* fix status for 2025h1 goals