Update design principles for reference goal to better reflect reference processes
Based on joint iteration by @traviscross and @joshtriplett. The new text addresses all the same concerns as the old text (notably making sure that we don't let the perfect become a blocker on the good), but lays out much more detailed processes, and avoids potentially incorrect or misleading implications about the normativity of the reference.

Co-authored-by: Travis Cross <tc@traviscross.com>