Expose reqwest's inner error in the `fetch` function (#424)
* Expose reqwest's inner error in the `fetch` function

We were converting the `reqwest::Error` type directly to `span::Error`
and the rendering chain stopped at that. This ignored the
inner (source) error which contained the most useful information to
figure out what went wrong.

This gets us from:

    Error: error: failed to fetch: error: error decoding response body

to:

    Error: error: failed to fetch: error: error decoding response body: missing field `discord` at line 16049 column 1

* Generate a spanned::Error directly from the inner JSON error

The code is shorter and less hacky. The end-user message looks much
better too.

* Use `to_string()` instead of `format!`

* Move the question mark out of `Ok(json_response?)`