Fix dark mode support for blog posts
The Rust blog has a light/dark mode switcher people are relying on.
Right now, generating a blog post and copying it as-is makes those
blog posts unreadable when dark mode is enabled.

Eric Huss fixed that manually for the September and October updates:

https://github.com/rust-lang/blog.rust-lang.org/pull/1731/commits/823ab6508099e1130e54a16bb548e4cc52a6e654

This is using that same fix at the source.