Separate "Goals looking for help" from "Other goal updates"
If the "Goals looking for help" section is empty, the "Other goal
updates" header would be shown on the same line.

You can see that in the October project update blog:

https://blog.rust-lang.org/2025/11/19/project-goals-update-october-2025/#goals-looking-for-help

This adds a line break between the two to force them to render on
separate pages.

NOTE: this only happens on the Rust Blog so it's likely better to fix
the CSS over there or something.

Treat this as a workaround for now.