Fix code rendering in reports
When a GitHub comment contains a code block, this would look really
weirdly in the final mdbeoo update report/blog.

This is because pulldown-cmark inserts spurious paragraphs inside HTML
code blocks.

For example this:

    <p>some text here</p>
    <pre><code>foo()

    bar()
    </code></pre>

Will end up being rendered as:

    <p>some text here</p>
    <pre><code>foo()

    <p>
    bar()
    </code></pre>

This is triggered by the first paragraph there. A workaround is to
insert a newline between the paragraph and the `<pre>` section:

    <p>some text here</p>

    <pre><code>foo()

    bar()
    </code></pre>

This will be treated as proper HTML and pulldown-cmark will not make
any changes to it.